add container inference to TestingGroup
We assume that only a singular type of grouped instances of Testing containers is created. Eg. there is not a mix of MasterNodeRegTestContainer and RegTestContainer in a single group.

This allows us to just infer the type being returned by the provided init method or fall back to MasterNode which is provided by the default method.