add report script
Since the build works in a matrix I have yet to find a clean solution to a concatenated output that can be used in subsequent job to produce a nice formatted message.

This solution simply adds a script that creates QOL tags in the same job rather than relying on previous jobs.