feat(whale-api): add name on poolpairs api response (#1740)
<!--  Thanks for sending a pull request! -->

#### What this PR does / why we need it:
On the redesigned screens for Token Selection (swap), we need to display
the token name. In this PR, token name is added to the response of
poolpairs apis.

#### Which issue(s) does this PR fixes?:
<!--
(Optional) Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes #

#### Additional comments?:

- Affected endpoints: `/poolpairs`, `/poolpairs/paths/swappable/13`,
`/poolpairs/dexprices`. There can be more affected routes, so definitely
need the reviewers' help to check.
- For `/poolpairs` api, `name` is added inside `tokenA` and `tokenB`
data.

- Redesigned screen where we need to display the name:

![image](https://user-images.githubusercontent.com/17063281/188565549-018aea34-f280-465e-b6c4-1bed80812fae.png)